Start background implementation
I added background arrays and variables to keep track of the board and what the player has selected.
