+++ src/arcade/games/checkers/__init__.py
@@ -1,0 +1,5 @@
+
+from arcade.games.checkers.checkers import Checkers
+
+def insert_cartridge():
+    return Checkers()
